Create funding at bowieronin@gmail.com
That's our paypalaccount for the time being. We'll put to good use and eventually outthere again! For all the right reasons; to deploy, educate, express yourselves and creating a better World for everyone.
We will make assessments, informe you and ask for your opinion on things (subjects at hand) so you will be valued whatever you have to say.